Rename audit log folder
