+++ source/AuditLog/AuditLogClient/AuditLogOptions.cs
@@ -14,7 +14,7 @@
 
 using System.ComponentModel.DataAnnotations;
 
-namespace Energinet.DataHub.EDI.AuditLog.AuditLogServerClient;
+namespace Energinet.DataHub.EDI.AuditLog.AuditLogClient;
 
 /// <summary>
 /// Required configuration options for the <see cref="AuditLogHttpClient"/>.
